requirements.txt
streamlit
indic-nlp-library
moviepy
SpeechRecognition
spacy
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
numpy
scipy